### 2019-01-02 #1149 @whedon
[PRE REVIEW]: MIDI.jl: Intuitive handling of MIDI data made for musicians
**Submitting author:** @Datseris (<a href="http://orcid.org/0000-0002-6427-2385">George Datserus</a>)
**Repository:** <a href="https://github.com/JuliaMusic/MIDI.jl" target ="_blank">https://github.com/JuliaMusic/MIDI.jl</a>
**Version:** v1.0.0
