+++ paper/paper.bib
@@ -15,4 +15,4 @@ URL = {
 eprint = {
         https://doi.org/10.1137/141000671
 
-}
+}}
