### 2019-01-07 #1149 @Kevin-Mattheus-Moerman
@Datseris Please consult the [author guidelines](https://joss.readthedocs.io/en/latest/submitting.html) carefully to make sure your submission conforms to our standards. You can also check out the [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) to get an idea off what the reviewers will look for.

Can you work on the following:
- [ ] Please expand your readme
Templates:
https://gist.github.com/PurpleBooth/109311bb0361f32d87a2
https://github.com/dbader/readme-template

- [ ] Please add contributing guidelines. You can do so by adding a CONTRIBUTING.md file.
Guides:
https://help.github.com/articles/setting-guidelines-for-repository-contributors/
Example:
https://github.com/gibbonCode/GIBBON/blob/master/CONTRIBUTING.md
- [ ] Also consider adding a Code of Conduct document e.g. in the form of a COC.md file.
Base one on this template: https://www.contributor-covenant.org/
- [ ] Link to the CONTRIBUTING and COC files in the readme.
- [ ] I recommend renaming the LICENSE.md file to a LICENSE file and to remove the line `The MIDI.jl package is licensed under the MIT "Expat" License`.
