+++ CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# Contribution Guidelines
+
+## Reporting issues
+
+- **Search for existing issues.** Please check to see if someone else has reported the same issue.
+- **Share as much information as possible.** Include operating system and version, browser and version. Also, include steps to reproduce the bug.
+
+## Project Setup
+Refer to the [README](README.md).
+
+## Code Style
+
+Please follow the [official Julia style](https://docs.julialang.org/en/v1/manual/style-guide/index.html) guide throughout.
+
+## Testing
+All contributed code should be tested as any other Julia package. See the `test` folder for example tests.
+
+## Pull requests
+- Try not to pollute your pull request with unintended changes – keep them simple and small. You are not required to squash your commits and we do not encourage it.
+- If your PR resolves an issue, include **closes #ISSUE_NUMBER** in your commit message (or a [synonym](https://help.github.com/articles/closing-issues-via-commit-messages)).
+- Review
+    - If your PR is ready for review, another contributor will be assigned to review your PR. If one has not been assigned, please ping `@Datseris`.
+    - If needed address the comments left by the reviewer. Once you're ready to continue the review, ping the reviewer in a comment.
+    - Once accepted your code will be merged.
