### 2019-01-07 586695a
update readme as per request of JOSS standards
