### 2019-01-07 #1149 @Datseris
@Kevin-Mattheus-Moerman thank you for accepting to be the assigned editor. I have made all the changes that you requested, please tick them if you agree with the current `master` branch.

---

@ssfrr thank you very much for accepting to be a reviewer.
