### 2019-01-07 #1149 @Kevin-Mattheus-Moerman
@Datseris can you work on your paper style and content. Please refer to the JOSS [author guidelines](https://joss.readthedocs.io/en/latest/submitting.html).

- [ ] One of the requirements for JOSS is that the paper should have a clear research application. Can you please expand on this in the paper? What scientific/research work does this software target? Currently the paper boasts "made for musicians" and "not computer scientists or engineers". Perhaps rephrase those parts. It may be that the intuitive/clear way in which your work handles MIDI files is very useful in research. Can you cite research to which this work would apply?
- [ ] "This is very handful,..." should probably be "This is very handy,..." or consider rephrasing to "This is very convenient,...".
- [ ] "like e.g." is double, choose e.g., like, or for example.

Let me know if I can help or if you have questions.
