### 2019-01-08 #1149 @Datseris
Thank you for your comments @Kevin-Mattheus-Moerman . I have done everything you suggested besides your first point on research applications.

I want to quickly state that this "boasting" that you mention did not come from myself. Rather, users of MIDI.jl contacted me to tell how much easier this has made their work and how it is so much easier to use than other MIDI packages. I now realize that such a fact can only be understood by someone that has worked with MIDI data and has used MIDI.jl. Therefore I will fully rephrase these parts of the paper.

As far as research applications are concerned, the package has plentiful. In fact we just submitted a scientific paper for review in a well respected journal. I hope we have it soon on arXiv so I can cite it. (In fact this is the reason I am submitting this paper, so I can cite it in the main paper)

Even if we don't have the paper on arXiv on time, there are plenty of examples I can show that MIDI.jl can be used in scientific research. I will add a corresponding paragraph as soon as possible.
