### 2019-01-08 #1149 @Kevin-Mattheus-Moerman
@jfsantos would you be interested in reviewing this submission for the [Journal of Open Source Software (JOSS)](http://joss.theoj.org/). The review process is focused on evaluating [the software](https://github.com/JuliaMusic/MIDI.jl) and [a short paper](https://github.com/openjournals/joss-papers/blob/joss.01149/joss.01149/10.21105.joss.01149.pdf) ([here is an example review process](https://github.com/openjournals/joss-reviews/issues/966) where you see a checkbox list guiding reviewers through the process). Let me know if you are interested or if you have questions.
