### 2019-01-08 771e5b2
full paper rework
