+++ paper/paper.bib
@@ -16,3 +16,818 @@ eprint = {
         https://doi.org/10.1137/141000671
 
 }}
+
+% Encoding: UTF-8
+
+@Article{Nye1937,
+  author    = {Nye, Russel B},
+  title     = {A musician's word list},
+  journal   = {American Speech},
+  year      = {1937},
+  volume    = {12},
+  number    = {1},
+  pages     = {45--48},
+  file      = {:Nye1937 - A musician's word list.pdf:PDF},
+  publisher = {JSTOR},
+}
+
+@Article{Levitin2018,
+  author   = {Levitin, Daniel J and Grahn, Jessica A and London, Justin},
+  title    = {The Psychology of Music: Rhythm and Movement},
+  journal  = {Annu. Rev. Psychol.},
+  year     = {2018},
+  volume   = {69},
+  number    = {1},
+  pages     = {51--75},
+  abstract = {The urge to move to music is universal among humans. Unlike visual art,
+which is manifest across space, music is manifest across time. When listeners
+get carried away by the music, either through movement (such as dancing)
+or through reverie (such as trance), it is usually the temporal qualities of
+the music—its pulse, tempo, and rhythmic patterns—that put them in this
+state. In this article, we review studies addressing rhythm, meter, movement,
+synchronization, entrainment, the perception of groove, and other tempo-
+ral factors that constitute a ﬁrst step to understanding how and why music
+literally moves us. The experiments we review span a range of method-
+ological techniques, including neuroimaging, psychophysics, and traditional
+behavioral experiments, and we also summarize the current studies of animal
+synchronization, engaging an evolutionary perspective on human rhythmic
+perception and cognition.},
+  file     = {:Levitin2018 - The Psychology of Music_ Rhythm and Movement.pdf:PDF},
+  groups   = {perception analysis, introduction, Entrainment},
+  keywords = {rank5},
+}
+
+@Article{Proegler1995,
+  author    = {J. A. Pr{\"{o}}gler},
+  title     = {Searching for Swing: Participatory Discrepancies in the Jazz Rhythm Section},
+  journal   = {Ethnomusicology},
+  year      = {1995},
+  volume    = {39},
+  number    = {1},
+  pages     = {21--54},
+  file      = {:Proegler1995 - Searching for Swing_ Participatory Discrepancies in the Jazz Rhythm Section.pdf:PDF},
+  publisher = {University of Illinois Press},
+  timestamp = {2018-04-22},
+}
+
+@Article{Janata2012,
+  author    = {Janata, Petr and Tomic, Stefan T and Haberman, Jason M},
+  title     = {Sensorimotor coupling in music and the psychology of the groove.},
+  journal   = {J. Exp. Psychol. Gen.},
+  year      = {2012},
+  volume    = {141},
+  number    = {1},
+  pages     = {54},
+  file      = {:Janata2012 - Sensorimotor coupling in music and the psychology of the groove..pdf:PDF},
+  publisher = {US: American Psychological Association},
+}
+
+@Book{Apel2003,
+  title     = {The Harvard Dictionary of Music: Fourth Edition},
+  publisher = {Harvard University Press},
+  year      = {2003},
+  author    = {Apel, W. and Randel, D.M.},
+  series    = {Harvard University Press reference library},
+  isbn      = {9780674011632},
+  lccn      = {2003058262},
+}
+
+@Article{Madison2011,
+  author    = {Madison, Guy and Gouyon, Fabien and Ull{\'e}n, Fredrik and H{\"o}rnstr{\"o}m, Kalle},
+  title     = {Modeling the tendency for music to induce movement in humans: first correlations with low-level audio descriptors across music genres.},
[…]
