### 2019-01-08 #1149 @Datseris
@Kevin-Mattheus-Moerman I have reworked my manuscript and I believe that it satisfies the author guidelines of JOSS and provides a clear scientific application and further use.
