### 2019-01-12 #1166 @whedon
[REVIEW]: MIDI.jl: Intuitive handling of MIDI data made for musicians
**Submitting author:** @Datseris (<a href="http://orcid.org/0000-0002-6427-2385">George Datserus</a>)
**Repository:** <a href="https://github.com/JuliaMusic/MIDI.jl" target ="_blank">https://github.com/JuliaMusic/MIDI.jl</a>
**Version:** v1.1.3
