### 2019-01-12 #1166 @Kevin-Mattheus-Moerman
@Datseris, @ssfrr, @jfsantos, this is where the review takes place. There are instructions at the top of this issue as well as a check list which will guide you through the process. If all boxes are ticked and reviewers recommend acceptance this submission will be accepted in JOSS.

You can comment on minor things here and we encourage your to use check lists (`- [ ]` in markdown). Larger/longer issues are better discussed in a separate issue which you can open on the [submission repository](https://github.com/JuliaMusic/MIDI.jl) and refer to here.

Thanks again for helping with the review process!!!!!

Let me know if you have any questions.
