### 2019-01-12 #1166 @Kevin-Mattheus-Moerman
@Datseris can you work on fixing that figure in the paper? Let me know if you need help.
