### 2019-01-12 #1166 @Datseris
@Kevin-Mattheus-Moerman I do not know how to fix this figure, as I said in a previous comment in the previous issue. It is a standard `.png` figure that is included with the standard `![description](figure)` format. The figure doesn't have overly big whitespace either.

From my perspective this is a problem of the "compilation/pdf-generation" procedure. If you know what I have to do then please let me know.
