### 2019-01-12 #1166 @Kevin-Mattheus-Moerman
@arfon any idea what is causing the figure to be off center in [this paper](https://github.com/openjournals/joss-papers/blob/joss.01166/joss.01166/10.21105.joss.01166.pdf)?
