### 2019-01-12 7f967ba
Added empty lines around figure in paper
To help resolve issue in displaced figure in the JOSS paper
