+++ paper/paper.md
@@ -100,7 +100,9 @@ mtds_ms = mtds .* ms_per_tick(midi)
   13.461524999999998
 ```
 Plotting the histogram of these gives
+
 ![histogram](mtd_hist.png)
+
 which, even if produced with an extremely small pool of data, follows the existing evidence that the distribution of the microtiming deviations follows a normal distribution.
 
 # Necessity of MIDI.jl
