### 2019-01-12 #1166 @Kevin-Mattheus-Moerman
@Datseris looks a bit like it is an inline picture so I tried to add extra lines in the PR https://github.com/JuliaMusic/MIDI.jl/pull/111.
