### 2019-01-12 3b1d94c
update figure caption
