### 2019-01-20 c406e9c
typos in paper
