+++ paper/paper.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 @article{Julia,
 author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
 title = {Julia: A Fresh Approach to Numerical Computing},
@@ -17,7 +18,11 @@ eprint = {
 
 }}
 
-% Encoding: UTF-8
+@Article{Datseris2019,
+  author    = {Datseris, George and Ziereis, Annika and Albrecht, Thorsten and Hagmayer, York and Priesemann, Viola and Geisel, Theo},
+  title     = {Does it Swing? Microtiming Deviations and Swing Feeling in Jazz},
+  journal   = {to be publ.},
+}
 
 @Article{Nye1937,
   author    = {Nye, Russel B},
