### 2019-01-24 c1bca91
properly link figure
