### 2019-01-25 3bd3824
abide to scientific note notation (#112)
* abide to scientific note notation

* fix tests
