### 2019-01-25 c84f097
re-work when warning is shown
