### 2019-01-25 #1166 @Datseris
Hello @Kevin-Mattheus-Moerman , @jfsantos , @ssfrr . While I was trying to improve MIDI.jl I realized that I had to do a breaking bugfix. If you are interested you may look at https://github.com/JuliaMusic/MIDI.jl/pull/112 but this is in no way relevant to the paper.

Although this is a breaking change it is a bugfix and thus does not require a major version increment. Regardless, I have released version 1.1 that includes a warning of this change. if possible, please associate this paper with version 1.1 and not with version 1.0. Thank you.
