### 2019-01-25 2fda46a
prettyprint vector of notes as well
