### 2019-01-25 8a8ea1b
big increase of note vector display
