### 2019-01-26 66eb2fd
prettyprint vector of notes as well (#113)
* prettyprint vector of notes as well

* big increase of note vector display
