### 2019-02-15 #1166 @jfsantos
Hello all, I just started my review. The version I get from Github is 1.1.0 but the review form says 1.0.0. The paper doesn't mention a specific version, so I guess that is a non-issue?
