### 2019-02-15 #1166 @jfsantos
Other minor issues:

- [ ] Most of the references are missing a DOI.
- [ ] Figure 1 is mentioned in the text as appearing right below the text, but is actually rendered on the next page. Please fix the text and point the reader to Figure 1 instead of assuming it is going to be rendered on a specific position.
