### 2019-02-15 #1166 @jfsantos
Another question:

- The paper only describe examples of note events on MIDI, but the protocol/format deals with a lot more than just note events (like velocity changes, communication with multiple instruments in different channels, etc). Does the software address these other events? Is it future work or do the maintainers have no interest in supporting those?
