### 2019-02-15 #1166 @Datseris
Thank you for the review @jfsantos .

> Figure 1 is mentioned in the text as appearing right below the text, but is actually rendered on the next page. Please fix the text and point the reader to Figure 1 instead of assuming it is going to be rendered on a specific position

Please read the latest version of the pdf, where this issue is fixed. whendon will compile a new one for you now. Towards my side there is only two issues: remove references with missing DOIs and answer the question:

> The paper only describe examples of note events on MIDI, but the protocol/format deals with a lot more than just note events (like velocity changes, communication with multiple instruments in different channels, etc). Does the software address these other events? Is it future work or do the maintainers have no interest in supporting those?

The core MIDI protocol is already implemented in MIDI.jl. These details far exceed the scope of a publication and are not discussed in the paper (and I believe I am already at the limit of word count suggested by JOSS).

Please read the documentation for more: in short, all MIDI/META/SysEx events are possible to write and read using MIDI.jl as well as any instrument channel (which as a result covers the entire MIDI protocol).

Anything exceeding what is already implemented is not of interest for the developer of MIDI.jl.
