### 2019-02-15 #1166 @jfsantos
> Please read the documentation for more: in short, all MIDI/META/SysEx events are possible to write and read using MIDI.jl as well as any instrument channel (which as a result covers the entire MIDI protocol).

Maybe just mention this somewhere in the paper or documentation? I tried to find it but on a quick search did not find anything mentioning events that were not notes.

> Anything exceeding what is already implemented is not of interest for the developers of MIDI.jl.

Sounds good, and I don't think this needs to be mentioned anywhere.
