### 2019-02-15 #1166 @Datseris
> Maybe just mention this somewhere in the paper or documentation? I tried to find it but on a quick search did not find anything mentioning events that were not notes.

I am not sure I agree with this statement, because this is already mentioned in the documentation. Here is my justification:

Searching the documentation (using the search bar) for "META EVENT" brings me here: https://juliamusic.github.io/JuliaMusic_documentation.jl/dev/midi/io/#MIDI.TrackEvent which lists clearly (at least I believe so) the existence of and how to create or read META/MIDI/SysEx events.

If one can create any of these three track events then one can implement the entire midi protocol.

---

Please let me know if this is still unclear.
