+++ paper/paper.bib
@@ -24,43 +24,6 @@ eprint = {
   journal   = {to be publ.},
 }
 
-@Article{Nye1937,
-  author    = {Nye, Russel B},
-  title     = {A musician's word list},
-  journal   = {American Speech},
-  year      = {1937},
-  volume    = {12},
-  number    = {1},
-  pages     = {45--48},
-  file      = {:Nye1937 - A musician's word list.pdf:PDF},
-  publisher = {JSTOR},
-}
-
-@Article{Levitin2018,
-  author   = {Levitin, Daniel J and Grahn, Jessica A and London, Justin},
-  title    = {The Psychology of Music: Rhythm and Movement},
-  journal  = {Annu. Rev. Psychol.},
-  year     = {2018},
-  volume   = {69},
-  number    = {1},
-  pages     = {51--75},
-  abstract = {The urge to move to music is universal among humans. Unlike visual art,
-which is manifest across space, music is manifest across time. When listeners
-get carried away by the music, either through movement (such as dancing)
-or through reverie (such as trance), it is usually the temporal qualities of
-the music—its pulse, tempo, and rhythmic patterns—that put them in this
-state. In this article, we review studies addressing rhythm, meter, movement,
-synchronization, entrainment, the perception of groove, and other tempo-
-ral factors that constitute a ﬁrst step to understanding how and why music
-literally moves us. The experiments we review span a range of method-
-ological techniques, including neuroimaging, psychophysics, and traditional
-behavioral experiments, and we also summarize the current studies of animal
-synchronization, engaging an evolutionary perspective on human rhythmic
-perception and cognition.},
-  file     = {:Levitin2018 - The Psychology of Music_ Rhythm and Movement.pdf:PDF},
-  groups   = {perception analysis, introduction, Entrainment},
-  keywords = {rank5},
-}
 
 @Article{Proegler1995,
   author    = {J. A. Pr{\"{o}}gler},
@@ -110,6 +73,7 @@ Keywords: adjective ratings; audio analysis; coordination; computational modelli
   file      = {:Madison2011 - Modeling the tendency for music to induce movement in humans_ first correlations with low-level audio descriptors across music genres..pdf:PDF},
   publisher = {American Psychological Association},
   timestamp = {2018-04-22},
+  doi = {10.1037/a0024323},
 }
 
 @Article{Butterfield2010,
@@ -125,41 +89,9 @@ Keywords: adjective ratings; audio analysis; coordination; computational modelli
   file      = {:butterfield_2009_participatory_discrepancies_and_the_perception_of_beats_in_jazz.pdf:PDF},
   publisher = {University of California Press Journals},
   timestamp = {2017-03-12},
+  doi = {10.1525/mp.2010.27.3.157},
 }
 
-@Article{Madison2014,
-  author    = {Madison, Guy and Sioros, George},
-  title     = {What musicians do to induce the sensation of groove in simple and complex melodies, and how listeners perceive it},
-  journal   = {Front. Psychol.},
-  year      = {2014},
-  volume    = {5},
-  pages     = {894},
-  abstract  = {Groove is the experience of wanting to move when hearing music, such as snapping
-fingers or tapping feet. This is a central aspect of much music, in particular of music
-intended for dancing. While previous research has found considerable consistency in
-ratings of groove across individuals, it remains unclear how groove is induced, that is,
-what are the physical properties of the acoustic signal that differ between more and less
-groove-inducing versions. Here, we examined this issue with a performance experiment,
-in which four musicians performed six simple and six complex melodies in two conditions
-with the intention of minimizing and maximizing groove. Analyses of rhythmical and
-temporal properties from the performances demonstrated some general effects. For
[…]
