### 2019-02-15 #1166 @Datseris
Alright, I have fixed the missing DOIs. Now all cited literature includes a DOI.
