### 2019-02-15 aa97797
add citation badge
