### 2019-02-15 #1166 @jfsantos
I updated the checkboxes for the minor issues and review accordingly. I don't know what are the specific guidelines for versioning on JOSS but the paper does not describe which version of MIDI.jl it is talking about.
