### 2019-02-15 #1166 @Datseris
Ideally I would like this paper to be associated with version 1.1.0. I have added an explicit statement about this in the manuscript now. Of course, it falls into the hands of @Kevin-Mattheus-Moerman whether this will be accepted and further more implemented, as it falls beyond my power to make such a change on `whedon`.
