### 2019-02-18 #1166 @ssfrr
Getting rolling now, basing my review on the proof from [3 days ago](https://github.com/openjournals/joss-reviews/issues/1166#issuecomment-464238975)
