### 2019-02-18 #1166 @ssfrr
This paper also references and uses MusicManipulations.jl. Is that in-scope for the review, or are we solely considering MIDI.jl?
