### 2019-02-18 #1166 @ssfrr
## Review Summary:

MIDI.jl is a Julia package to enable users to read and write MIDI notes and tracks. While there is prior work in this space, it is useful to have a pure-julia implementation for working with MIDI files, as MIDI is a widely-used format for symbolic music representation. This work also includes the MusicManipulations package, which provides additional functionality for quantizing MIDI notes to a predefined grid, translating and transposing arrays of notes, filtering notes, and extracting individual properties into separate time series.

## Specific Issues

-   [ ] The paper makes the claim "This makes reading MIDI data easier than ever before.", which seems a little promotional and isn't substantiated.
-   [ ] There are multiple references to saving the user from worrying about understanding the byte-level representation of MIDI files. I think this can be assumed of any package that parses formatted data files and can be removed.
-   [ ] Most of the references are to papers on microtiming, which doesn't seem closely tied to MIDI.jl (though possibly to MusicManipulations.jl). I think these can be pared down to only what's relevant to the given example.
-   [ ] There's unreferenced prior art, e.g. [mido](https://github.com/mido/mido), [python-midi](https://github.com/vishnubob/python-midi), [pretty-midi](https://github.com/craffel/pretty-midi) and the [accompanying paper](https://colinraffel.com/publications/ismir2014intuitive.pdf). That paper also has some references to other packages (as of 2014). It would be useful to see some more discussion of how MIDI.jl compares in approach to these prior libraries. There's also [this MATLAB implementation](http://kenschutte.com/midi), but given the dependence on MATLAB and the relative complexity of calling into MATLAB relative to Python that is a less compelling comparison. The microtiming example could be slimmed down to compensate.
