### 2019-02-19 #1166 @Datseris
Thank you @ssfrrr for your review. Let me know answer one by one your specific issues:

1. Thank you for pointing this out, I have now reworked my introduction.
2. Please be more specific about this issue. By reading the entire paper I can only find a singular instance: the opening paragraph of the section "Intuitive and simple interface". I could not find any other instance where there are "references to saving the user from worrying about understanding the byte-level representation of MIDI files.". If you believe there are multiple instances of this event please do state them explicitly. In addition, this software (and thus the paper) does not target audio engineers but musicians and scientists. Therefore it cannot be assumed that our users will "just know that they don't have to know about low level representations". We believe this has to be explicitly stated once, and we do so in our paper, again only once.
3. The editor of this review, @Kevin-Mattheus-Moerman asked me explicitly to include a "clear research application" in our paper, see the comment here: https://github.com/openjournals/joss-reviews/issues/1149#issuecomment-452094559 . As a scientist actively doing research in this field and actively publishing papers in this field I strongly stand that the current number of references should not be reduced further as they are all fully relevant to the example scientific application.
4. Thank you very much for sharing these references. I have included all of them in the paper (the MATLAB implementation was disregarded) and I have also added notes that compare **MIDI.jl** to them, in a concluding paragraph.
