### 2019-02-22 #1166 @ssfrr
I updated the checklist of issues in [my comment above](https://github.com/openjournals/joss-reviews/issues/1166#issuecomment-464843735).

Please cite their paper as a real reference

Thank you for adding the comparison. Can you clarify items 1 and 2 in the "notable differences" list? Specifically:

> The Notes data structure and getnotes functionality that exists in MIDI.jl

Can you expand on how the data structure and note retrieval functionality are different in MIDI.jl from the other libraries? It seems that the main advantage over the other libraries is that MIDI.jl combines the note-on and note-off events into a single note object with a duration (which is definitely a useful feature of MIDI.jl), but pretty-midi seems to have a [similar representation](http://craffel.github.io/pretty-midi/#pretty-midi-note). If that's the main difference maybe clarify that here.

> The extensibility of MIDI.jl into higher level applications not limited only to raw midi data (see MusicManipulations.jl).

Here again pretty-midi seems to provide some high-level functionality, for instance tempo and beat estimation and note transition matrices.
