+++ paper/paper.md
@@ -115,11 +115,11 @@ There is existing software that offers functionality similar, but not identical,
 
 * [mido](https://github.com/mido/mido)
 * [python-midi](https://github.com/vishnubob/python-midi)
-* [pretty-midi](https://github.com/craffel/pretty-midi), described also in this [document](https://colinraffel.com/publications/ismir2014intuitive.pdf)
+* [pretty-midi](https://github.com/craffel/pretty-midi), which also has an associated publication [@RaffelMHSNLE14].
 
 Notable differences between **MIDI.jl** and these libraries include (but are not limited to):
-1. The `Notes` data structure and `getnotes` functionality that exists in **MIDI.jl** 
-2. The extensibility of **MIDI.jl** into higher level applications not limited only to raw midi data (see **MusicManipulations.jl**).
+1. The `Notes` data structure and `getnotes` functionality that exists in **MIDI.jl**. Something similar exists in `pretty-midi` as well but for example there is no `channel` property.
+2. **MIDI.jl** is extended further into higher level applications like the ones offered by **MusicManipulations.jl** or the module `MotifSequenceGenerator` that can create specially random sequences of notes.
 3. The fact that **MIDI.jl** is written for the Julia programming language.
 4. **MIDI.jl** has proper, multi-page and multi-example documentation that is hosted online and is automatically updated with every commit to the repository.
 4. **MIDI.jl** does not currently have the so-called "piano roll" functionality, which plots notes as in a DAW.
