### 2019-02-22 #1166 @Datseris
Thank you for your reply.

> Please cite their paper as a real reference

Done.

> Can you expand on how the data structure and note retrieval functionality are different in MIDI.jl from the other libraries?

For **MIDI.jl** the `Note` is a central structure used everywhere and expanded by other modules. It has a much more important place than just a utility structure. Many functions (as shown in our official documentation) dispatch directly on the `Note` structure. In addition it has the `channel` property which I could not find elsewhere. Other than that indeed the "pretty-midi" representation seems quite similar. The paper was updated accordingly.

> Here again pretty-midi seems to provide some high-level functionality, for instance tempo and beat estimation and note transition matrices.

I have changed the sentence to:

> 2. **MIDI.jl** is extended further into higher level applications like the ones offered by **MusicManipulations.jl** or the module `MotifSequenceGenerator` that can create specially random sequences of notes.

I believe that pretty-midi is not extended by neither **MusicManipulations.jl** nor `MotifSequenceGenerator` and thus this is a notable difference.

---

@Kevin-Mattheus-Moerman can you please resolve the conflict regarding the software version?
