### 2019-02-23 #1166 @ssfrr
The reference you have in your references section is for the wrong paper. Also the link you gave is incorrect, and gives a 404. They provide the correct citation as well as a link to the paper in their [README](https://github.com/craffel/pretty-midi).

Please be more careful with your references. As I'm sure you're aware, properly citing prior work is one of the more important aspects of academic publishing. It shouldn't take three rounds of review to get directly-relevant prior work cited.
