### 2019-02-23 #1166 @Datseris
> Please be more careful with your references. As I'm sure you're aware, properly citing prior work is one of the more important aspects of academic publishing. It shouldn't take three rounds of review to get directly-relevant prior work cited.

Wow, sorry for the mistake I did not realize this.

Unfortunately they do not provide the correct citation in their README as you claim. They provide a PDF file and a claim of publication. Of course I tried to find a DOI or an ISBN if the citation is for a book chapter. I have visited the official page for the conference the authors claim to have published in. It is here: https://dblp2.uni-trier.de/db/conf/ismir/ismir2014.html

As you will see for yourself, there is no paper titled "Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty_midi" (but there is a single paper by C. Raffel which I wrongly assumed the correct one). In addition, searching the entire database for author "Colin Raffel" again does not return the claimed paper. You can see this here: https://dblp2.uni-trier.de/pers/hd/r/Raffel:Colin

If you do know where this paper is published, please be kind and share it with us here. The "text" they provide in the README is not enough of a proof to create a proper BiBTeX entry as required per the JOSS standard.
