### 2019-02-24 #1166 @ssfrr
It appears that the short paper was accepted in the Late-Breaking/Demo category. The official conference website has a list of these accepted papers:

http://www.terasoft.com.tw/conf/ismir2014/AcceptedLBDPapers.html
