+++ paper/paper.bib
@@ -775,23 +775,15 @@ year = {2014}
     year      = {1946},
 }
 
-@inproceedings{RaffelMHSNLE14,
+@inproceedings{Raffel,
   author    = {Colin Raffel and
-               Brian McFee and
-               Eric J. Humphrey and
-               Justin Salamon and
-               Oriol Nieto and
-               Dawen Liang and
                Daniel P. W. Ellis},
-  title     = {MIR{\_}EVAL: {A} Transparent Implementation of Common {MIR} Metrics},
+  title     = {INTUITIVE ANALYSIS, CREATION AND MANIPULATION OF MIDIDATA WITHpretty_midi},
   booktitle = {Proceedings of the 15th International Society for Music Information
-               Retrieval Conference, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
+               Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
                2014},
-  pages     = {367--372},
   year      = {2014},
-  crossref  = {DBLP:conf/ismir/2014},
-  url       = {http://www.terasoft.com.tw/conf/ismir2014/proceedings/T066\_320\_Paper.pdf},
-  timestamp = {Fri, 19 Dec 2014 14:08:38 +0100},
-  biburl    = {https://dblp.org/rec/bib/conf/ismir/RaffelMHSNLE14},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
+  crossref  = {http://www.terasoft.com.tw/conf/ismir2014/AcceptedLBDPapers.html},
+  url       = {http://www.terasoft.com.tw/conf/ismir2014/LBD/LBD29.pdf},
+  bibsource = {www.terasoft.com}
 }
