### 2019-02-24 #1166 @Datseris
Thank you @ssfrr for helping me with this. I've now corrected the citation (which as a result also corrected the wrong link).

I have to note that even though I searched the website, I couldn't find an associated DOI for this paper (I don't think one exists). But I hope this is okay because there is at list a link that links to the paper.
