### 2019-02-25 #1166 @ssfrr
The title for the raffel citation isn't showing up in the references. Maybe something weird with the bibtex?
