+++ paper/paper.bib
@@ -778,7 +778,7 @@ year = {2014}
 @inproceedings{Raffel,
   author    = {Colin Raffel and
                Daniel P. W. Ellis},
-  title     = {INTUITIVE ANALYSIS, CREATION AND MANIPULATION OF MIDIDATA WITH pretty_midi},
+  title     = {{INTUITIVE ANALYSIS, CREATION AND MANIPULATION OF MIDIDATA WITH pretty_midi}},
   booktitle = {Proceedings of the 15th International Society for Music Information
                Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
                2014},
