+++ paper/paper.bib
@@ -775,11 +775,11 @@ year = {2014}
     year      = {1946},
 }
 
-@inproceedings{Raffel,
+@article{Raffel,
   author    = {Colin Raffel and
                Daniel P. W. Ellis},
-  title     = {{INTUITIVE ANALYSIS, CREATION AND MANIPULATION OF MIDIDATA WITH pretty_midi}},
-  booktitle = {Proceedings of the 15th International Society for Music Information
+  title     = {{Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty_midi}},
+  journal   = {Proceedings of the 15th International Society for Music Information
                Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
                2014},
   year      = {2014},
