+++ paper/paper.bib
@@ -776,8 +776,7 @@ year = {2014}
 }
 
 @article{Raffel,
-  author    = {Colin Raffel and
-               Daniel P. W. Ellis},
+  author    = {Raffel, Colin and Ellis, Daniel P. W.},
   title     = {{Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty_midi}},
   journal   = {Proceedings of the 15th International Society for Music Information
                Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
