### 2019-02-25 080f5f5
strip bibtex to basics
