+++ paper/paper.bib
@@ -782,7 +782,5 @@ year = {2014}
                Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
                2014},
   year      = {2014},
-  crossref  = {http://www.terasoft.com.tw/conf/ismir2014/AcceptedLBDPapers.html},
   url       = {http://www.terasoft.com.tw/conf/ismir2014/LBD/LBD29.pdf},
-  bibsource = {www.terasoft.com}
 }
