### 2019-02-25 #1166 @Datseris
Alright, no matter what I try the title of the paper does not appear in the compiled pdf. If anyone knows the solution please let me know as well. The Bib entry I have now (which I stripped to the bare basics) is:
```
@article{Raffel,
  author    = {Raffel, Colin and Ellis, Daniel P. W.},
  title     = {{Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty_midi}},
  journal   = {Proceedings of the 15th International Society for Music Information
               Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
               2014},
  year      = {2014},
  url       = {http://www.terasoft.com.tw/conf/ismir2014/LBD/LBD29.pdf},
}
```
