+++ paper/paper.bib
@@ -777,7 +777,7 @@ year = {2014}
 
 @article{Raffel,
   author    = {Raffel, Colin and Ellis, Daniel P. W.},
-  title     = {{Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty_midi}},
+  title     = {{Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty midi}},
   journal   = {Proceedings of the 15th International Society for Music Information
                Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
                2014},
