### 2019-02-25 #1166 @Datseris
Fiiiiiiiiiiiiiinaly. It is not possible to have underscore in names in latex apparently.
