### 2019-02-25 7094e9f
use Authors bibtex entry
