+++ paper/paper.bib
@@ -775,12 +775,10 @@ year = {2014}
     year      = {1946},
 }
 
-@article{Raffel,
-  author    = {Raffel, Colin and Ellis, Daniel P. W.},
-  title     = {{Intuitive Analysis, Creation and Manipulation of MIDI Data with pretty midi}},
-  journal   = {Proceedings of the 15th International Society for Music Information
-               Retrieval Conference Late Breaking and Demo Papers, {ISMIR} 2014, Taipei, Taiwan, October 27-31,
-               2014},
-  year      = {2014},
-  url       = {http://www.terasoft.com.tw/conf/ismir2014/LBD/LBD29.pdf},
+@inproceedings{Raffel,
+  title={Intuitive analysis, creation and manipulation of MIDI data with pretty\_midi},
+  author={Raffel, Colin and Ellis, Daniel PW},
+  booktitle={15th International Society for Music Information Retrieval Conference Late Breaking and Demo Papers},
+  pages={84--93},
+  year={2014}
 }
