### 2019-03-01 #1166 @Kevin-Mattheus-Moerman
@ssfrr, @jfsantos thanks for your review work!! Please can you check if you can tick those final boxes?

Note that the version will be updated before acceptance so that box (which has been removed in future review issues) can safely be ticked.

@ssfrr, can you check the reference which should now be in good shape?
