### 2019-03-01 #1166 @Kevin-Mattheus-Moerman
Some minor editorial comments on the paper

- [ ] Can  you check the sentence "A  music note can be (in its most basic level) deconstructed into just four numbers: the temporal position that the note is played, the duration, the pitch and the intensity (strength with which the note is played, also called velocity)". Perhaps it should read: "A music note can be (in its most basic level) deconstructed into just four numbers: the temporal position **~~that~~** the note is played **in**, the duration, the pitch[,] and the intensity (strength with which the note is played, also called velocity)"

- [ ] "This easy to use high level interface allows MIDI.jl to be extendable. In e.g. another
software package MusicManipulations.jl we provide general functions for manipulating
(and further analyzing) music data.
For example, the function quantize from the package MusicManipulations.jl allows
the user to quantize any Notes instance to any grid." I propose to remove "this", to rephrase the part with "e.g.", and to remove the enter/paragraph break. i.e.: -> "**The** easy to use high level interface allows MIDI.jl to be extendable. For instance ~~ In e.g~~, in another software package MusicManipulations.jl we provide general functions for manipulating (and further analyzing) music data. For example, the function quantize from the package MusicManipulations.jl allows the user to quantize any Notes instance to any grid."

- [ ] Check the use of "high level" and "high-level" please use the hyphen consistently

- [ ] Add comma after "In addition" in "In addition[,] it has plenty more use for scientific applications."

- [ ] Please remove: ", which also has an associated publication" (i.e. the citation alone is sufficient)

- [ ] Consider rephrasing the following: "Something similar exists in pretty-midi as well but for example there is no channel property." e.g. to something like "Although the package pretty-midi contains similar functionality, it lacks the channel property."

- [ ] If DAW is an acronym can you please define it?

- [ ] Consider rephrasing "There is also no sequencer functionality in MIDI.jl, which is in example in python-midi." to something along the lines of: "Sequencer functionality, which has been implemented in for instance the python-midi package, is currently lacking in MIDI.jl"

- [ ] The following sentence sounds a bit vague, can you rephrase or perhaps expand with more examples too: "pretty-midi has some other high level functionality (like e.g. basic tempo estimation) also absent in MIDI.jl.", perhaps this improves the sentence especially if you expand it a bit: "MIDI.jl currently lacks some basic high-level functionality such as tempo estimation, which have been implemented in packages like pretty-midi."

- [ ] For the reference "Does it swing? Microtiming deviations and swing feeling in jazz", I would recommend (if possible) to upload it as a pre-print on some pre-print server (if allowed by publisher see here: http://www.sherpa.ac.uk/romeo/index.php), and to cite the DOI of the pre-print in this paper.
