+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 **MIDI.jl** is a Julia [@Julia] package for reading, writing and analyzing [MIDI](https://www.midi.org/specifications) data. In this paper we are briefly overviewing versions `1.1.0` or later for **MIDI.jl**.
 
 MIDI (Music Instrument Digital Interface) is a data format created to transmit music data across devices and computers. The [actual MIDI interface](https://www.midi.org/specifications) is very low level, directly translating all music information to and from byte chunks.
-**MIDI.jl** exposes all this low level interface, but it also builds a useable high-level interface on top of that. This makes reading MIDI data intuitive and convenient, as we demonstrate in the following examples.
+**MIDI.jl** exposes all this low level interface, but it also builds a useable high level interface on top of that. This makes reading MIDI data intuitive and convenient, as we demonstrate in the following examples.
 
 # Documentation
 All functionality of **MIDI.jl** is very well documented and hosted online. The documentation link can be found [here](https://juliamusic.github.io/JuliaMusic_documentation.jl/latest/).
@@ -38,7 +38,7 @@ Because of the rich documentation, in this paper we will only showcase specific 
 # Intuitive and Simple Interface
 The biggest strength of **MIDI.jl** is the ability to transform the raw MIDI data in a format that is human-readable, intuitive and simple to use and manipulate. In addition the high level interface does not require knowledge of which exact MIDI code corresponds to which exact MIDI command.
 
-What makes this possible is the data structures we create in order to provide easier handling of MIDI files. The most important data structure is the `Note`/`Notes`. A music note can be (in its most basic level) deconstructed into just four numbers: the temporal position that the note is played, the duration, the pitch and the intensity (strength with which the note is played, also called velocity). A `Note` is a data structure that has these four "quantities" as its "fields". All of these are accessible immediately with e.g. `Note.position` and their values can be mutated in place.
+What makes this possible is the data structures we create in order to provide easier handling of MIDI files. The most important data structure is the `Note`/`Notes`. A music note can be (in its most basic level) deconstructed into just four numbers: the temporal position the note is played in, the duration, the pitch, and the intensity (strength with which the note is played, also called velocity). A `Note` is a data structure that has these four "quantities" as its "fields". All of these are accessible immediately with e.g. `Note.position` and their values can be mutated in place.
 
 These aspects can be deduced from the raw MIDI format with a lot of analyzing of bytes. However, in **MIDI.jl** we provide a simple function:
 ```julia
@@ -47,9 +47,10 @@ getnotes(midi, args...)
 which obtains all note-specific information and stores it as a vector of notes, which we call `Notes`. This is very convenient, as to "identify" a note in the MIDI format, one needs to first identify two different streams of bytes; one denotes the start and the other the end of the note. This quickly becomes tedious, but `getnotes` does not expose all these details to the user.
 
 # Extensions
-This easy to use high level interface allows **MIDI.jl** to be extendable. In e.g. another software package **MusicManipulations.jl** we provide general functions for manipulating (and further analyzing) music data.
+The easy to use high level interface allows **MIDI.jl** to be extendable.  For instance, in another software package **MusicManipulations.jl** we provide general functions for manipulating (and further analyzing) music data.
+For example, the function `quantize` from the package **MusicManipulations.jl** allows the user to quantize any `Notes` instance to any grid.
 
[…]
