### 2019-03-01 #1166 @Datseris
@Kevin-Mattheus-Moerman thank you for your reply.

I have implemented all your suggestions besides the final one. Unfortunately the research group associated with this publication has decided to _not_ publish it in a non peer-reviewed outlet (such as pre-print servers). In addition the paper is currently under review in a peer-reviewed journal, which makes me believe there is no reason to waste resources into putting it on some archive.

But there is no problem. If JOSS does not allow citing to-be-published work I can simply remove the paragraph that references this work. I believe the rest of the manuscript already makes a strong case for the scientific relevance of **MIDI.jl**.
