### 2019-03-01 #1166 @Kevin-Mattheus-Moerman
@Datseris I've gone through your revisions. There is one more point remaining:

- [ ] Please remove: ", which also has an associated publication" (i.e. the citation alone is sufficient)
