### 2019-03-01 #1166 @Datseris
> That is fine. The pre-print is only a recommendation so you can leave it as is. Thanks.

Thank you very much for understanding.

> Please remove: ", which also has an associated publication" (i.e. the citation alone is sufficient)

Ops, so sorry I forgot this. It is fixed now!
