### 2019-03-12 #1166 @Datseris
@Kevin-Mattheus-Moerman am I supposed to do something? From the compiled pdf I can see that all articles that do have a DOI display it properly. I think whedon is pointing out references that I have in the `.bib` file but I don't actively use in the article.
