### 2019-03-12 #1166 @Kevin-Mattheus-Moerman
@Datseris okay. Thanks for clarifying that. If you clean up the .bib file we can avoid this error but it sounds like your paper is in good shape.
