### 2019-03-12 #1166 @Kevin-Mattheus-Moerman
@Datseris looks like we are nearly there. I will soon recommend that this work is accepted in JOSS. At this point, can you please:

- [ ] Archive a copy of the reviewed software on a service like [Zenodo](https://zenodo.org/) (some follow [these steps](https://guides.github.com/activities/citable-code/)) and list the DOI of this archived version here?

- [ ] Provide the version/release tag for the reviewed and archived version? E.g. it likely moved on from v1.0.0

Thanks
