### 2019-03-12 #1166 @Datseris
Thanks for the support @Kevin-Mattheus-Moerman , here is the DOI from ZENODO:

[![DOI](https://zenodo.org/badge/41460547.svg)](https://zenodo.org/badge/latestdoi/41460547)

The version is `v1.1.3`
