### 2019-03-12 #1166 @labarba
I do have some editorial suggestions on the paper:

- Documentation: "very well documented" >> remove "very"
- "The documentation link can be found here." >> It's awkward to have "here" as an active link in a PDF (or even a website, where it's considered bad form!). You could add the URL to the references list, instead? Or simply spell out the full URL (or create a short URL) in the body text.
- "Because of the rich documentation, in tis paper…" >> this sentence is unnecessary, because all JOSS papers are requested _not_ to include any documentation.
- Given the above notes, I recommend removing the heading "Documentation" and merging the sentences left to the "Introduction" section.
- various places: high-level interface, low-level interface missing the hyphen in the compound adjective
- various missing commas, after "In this paper," after "In addition," after "e.g.," and "In the following."
- "the data structure we create" >> unclear: is it the data structure _you_ (the authors) created as part of your design process? (in which case it should be in past sense), or is ti the data structure that the software creates when it is being used? (in which case the subject of the sentence is not "we")
- "The easy to use high level interface" >> missing hyphens in compound adjectives
- Scientific Application: "see (references)">> if you are citing references as parts of a sentence, please use the Author (Year) format:  For the syntax to obtain brackets in the right places, see:
https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.htm
- Necessity of MIDI.jl — the fact that no other package offers the functionally does not make it necessary! I think you can remove the heading here, and move that lonely sentence into the Related Software section. (The case for need of the software was made in the previous section.)
- Do a search for the word "very" and in each appearance ask yourself if you really need it.

Finally, you are citing _"Does it swing? Microtiming deviations and swing feeling in jazz"_, which is apparently unpublished. Is it possible to deposit the preprint somewhere? (Like OSF preprints, or similar.) Citation to a non-public artifact is discouraged.
