+++ paper/paper.md
@@ -25,20 +25,18 @@ bibliography: paper.bib
 
 
 # Introduction
-**MIDI.jl** is a Julia [@Julia] package for reading, writing and analyzing [MIDI](https://www.midi.org/specifications) data. In this paper we are briefly overviewing versions `1.1.0` or later for **MIDI.jl**.
+**MIDI.jl** is a Julia @Julia package for reading, writing and analyzing [MIDI](https://www.midi.org/specifications) data. In this paper, we are briefly overviewing versions `1.1.0` or later for **MIDI.jl**.
 
-MIDI (Music Instrument Digital Interface) is a data format created to transmit music data across devices and computers. The [actual MIDI interface](https://www.midi.org/specifications) is very low level, directly translating all music information to and from byte chunks.
-**MIDI.jl** exposes all this low level interface, but it also builds a useable high level interface on top of that. This makes reading MIDI data intuitive and convenient, as we demonstrate in the following examples.
+MIDI (Music Instrument Digital Interface) is a data format created to transmit music data across devices and computers. The [actual MIDI interface](https://www.midi.org/specifications) is low-level, directly translating all music information to and from byte chunks.
+**MIDI.jl** exposes all this low-level interface, but it also builds a useable high-level interface on top of that. This makes reading MIDI data intuitive and convenient, as we demonstrate in the following examples.
 
-# Documentation
-All functionality of **MIDI.jl** is very well documented and hosted online. The documentation link can be found [here](https://juliamusic.github.io/JuliaMusic_documentation.jl/latest/).
+All functionality of **MIDI.jl** is well documented and hosted online: https://juliamusic.github.io/JuliaMusic_documentation.jl/latest/.
 Besides documentation of functions there are plenty of useful real world examples.
-Because of the rich documentation, in this paper we will only showcase specific strengths and design decisions that can make working with music data very easy.
 
 # Intuitive and Simple Interface
-The biggest strength of **MIDI.jl** is the ability to transform the raw MIDI data in a format that is human-readable, intuitive and simple to use and manipulate. In addition the high level interface does not require knowledge of which exact MIDI code corresponds to which exact MIDI command.
+The biggest strength of **MIDI.jl** is the ability to transform the raw MIDI data in a format that is human-readable, intuitive and simple to use and manipulate. In addition, the high-level interface does not require knowledge of which exact MIDI code corresponds to which exact MIDI command.
 
-What makes this possible is the data structures we create in order to provide easier handling of MIDI files. The most important data structure is the `Note`/`Notes`. A music note can be (in its most basic level) deconstructed into just four numbers: the temporal position the note is played in, the duration, the pitch, and the intensity (strength with which the note is played, also called velocity). A `Note` is a data structure that has these four "quantities" as its "fields". All of these are accessible immediately with e.g. `Note.position` and their values can be mutated in place.
+What makes this possible is the data structures we have designed in order to provide easier handling of MIDI files. The most important data structure is the `Note`/`Notes`. A music note can be (in its most basic level) deconstructed into just four numbers: the temporal position the note is played in, the duration, the pitch, and the intensity (strength with which the note is played, also called velocity). A `Note` is a data structure that has these four "quantities" as its "fields". All of these are accessible immediately with e.g., `Note.position` and their values can be mutated in place.
 
 These aspects can be deduced from the raw MIDI format with a lot of analyzing of bytes. However, in **MIDI.jl** we provide a simple function:
 ```julia
@@ -47,18 +45,18 @@ getnotes(midi, args...)
[…]
