### 2019-03-13 #1166 @Datseris
Hi @labarba ,

In the latest master commit I implemented all your points.

> Scientific Application: "see (references)">> if you are citing references as parts of a sentence, please use the Author (Year) format: For the syntax to obtain brackets in the right places, see:
https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.htm

I don't understand where you are pointing at... In text I have:

```
These have been studied extensively in the literature and their importance and influence are debated strongly, see  [@Madison2011, @Butterfield2010, @Fruehauf2013, @Davies2013, @Senn2016, @Hofmann2017] and references therein.
```

every reference is named. In the pdf all references are expanded to the Author (Year) format.

---

> Finally, you are citing "Does it swing? Microtiming deviations and swing feeling in jazz", which is apparently unpublished. Is it possible to deposit the preprint somewhere? (Like OSF preprints, or similar.) Citation to a non-public artifact is discouraged.

No, this is not possible, please see the discussion I had with @Kevin-Mattheus-Moerman  in the comment above:  https://github.com/openjournals/joss-reviews/issues/1166#issuecomment-468617973

> Citation to a non-public artifact is discouraged.

That may be true, but our group discourages non peer reviewed publications. We would appreciate your respecting of our choice.
