### 2019-03-13 #1166 @labarba
Regarding the citations, whether in-text or not, imagine if the format of the article used superscripts for citations: it would be quite weird to see a verb with a superscript but no subject, right? So, when you use the citation as part of speech in a sentence, you display it differently than if the reference is parenthetical (and could be replaced with a footnote and not look weird).

See: https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.htm#citation_syntax
... the part about in-text citations.
