### 2019-03-13 #1166 @labarba
Regarding your decision _not_ to post a pre-print of the other manuscript, that is of course your prerogative. But in that case, please remove it from the references list. You can still mention it in the text, with something like: "we currently have a manuscript under review titled "Blah" that used this software."

I'm still a little surprised by your answer, though. Pre-prints are increasingly recognized as a valid and important tool for open science. Most journals now accept pre-printing and don't count it as prior publication. Have a look at: https://en.wikipedia.org/wiki/List_of_academic_journals_by_preprint_policy
