### 2019-03-13 #1166 @Datseris
Hi @labarba ,

In the latest version I've tried to answer your comments again. By reading the "in-text citations" part what I've taken is that I should replace brackets with parenthesis.
