### 2019-03-13 #1166 @labarba
The documentation says:

You can also write an in-text citation, as follows:

`@smith04 says blah.`

No brackets.
