### 2019-03-13 #1166 @Datseris
> Looks like this is still pending?

Yeap, fixed now. I had to publish the changes: https://zenodo.org/record/2591437

> No brackets.

I've removed all brackets from current version. Let's see:
