### 2019-03-13 #1166 @labarba
It looks like you have a bunch of entries in the `paper.bib` file that are not cited in the paper. I think you will need to clean up the `.bib` file, because these are listed as citations in the `xml` that we send to Crossref. See:
https://github.com/openjournals/joss-papers/blob/6d4ea9c3a81a9603b6532c3eb9466cc9b93f3d42/joss.01166/10.21105.joss.01166.crossref.xml

@arfon — Can you confirm that we need the `paper.bib` file to contain _only_ the references cited in the text? I'm not sure about this.
