### 2019-03-13 #1166 @arfon
> @arfon — Can you confirm that we need the paper.bib file to contain only the references cited in the text? I'm not sure about this.

Yes please.
