+++ paper/paper.bib
@@ -18,47 +18,6 @@ eprint = {
 
 }}
 
-@Article{Datseris2019,
-  author    = {Datseris, George and Ziereis, Annika and Albrecht, Thorsten and Hagmayer, York and Priesemann, Viola and Geisel, Theo},
-  title     = {Does it Swing? Microtiming Deviations and Swing Feeling in Jazz},
-  journal   = {to be publ.},
-}
-
-
-@Article{Proegler1995,
-  author    = {J. A. Pr{\"{o}}gler},
-  title     = {Searching for Swing: Participatory Discrepancies in the Jazz Rhythm Section},
-  journal   = {Ethnomusicology},
-  year      = {1995},
-  volume    = {39},
-  number    = {1},
-  pages     = {21--54},
-  file      = {:Proegler1995 - Searching for Swing_ Participatory Discrepancies in the Jazz Rhythm Section.pdf:PDF},
-  publisher = {University of Illinois Press},
-  timestamp = {2018-04-22},
-}
-
-@Article{Janata2012,
-  author    = {Janata, Petr and Tomic, Stefan T and Haberman, Jason M},
-  title     = {Sensorimotor coupling in music and the psychology of the groove.},
-  journal   = {J. Exp. Psychol. Gen.},
-  year      = {2012},
-  volume    = {141},
-  number    = {1},
-  pages     = {54},
-  file      = {:Janata2012 - Sensorimotor coupling in music and the psychology of the groove..pdf:PDF},
-  publisher = {US: American Psychological Association},
-}
-
-@Book{Apel2003,
-  title     = {The Harvard Dictionary of Music: Fourth Edition},
-  publisher = {Harvard University Press},
-  year      = {2003},
-  author    = {Apel, W. and Randel, D.M.},
-  series    = {Harvard University Press reference library},
-  isbn      = {9780674011632},
-  lccn      = {2003058262},
-}
 
 @Article{Madison2011,
   author    = {Madison, Guy and Gouyon, Fabien and Ull{\'e}n, Fredrik and H{\"o}rnstr{\"o}m, Kalle},
@@ -145,23 +104,6 @@ Keywords: microtiming, groove, entrainment, body movement, participatory discrep
   doi = {10.3389/fpsyg.2016.01487},
 }
 
-@Article{Matsushita2016,
-  author    = {Matsushita, Soyogu and Nomura, Shingo},
-  title     = {The Asymmetrical Influence of Timing Asynchrony of Bass Guitar and Drum Sounds on Groove},
-  journal   = {Music Perception: An Interdisciplinary Journal},
-  year      = {2016},
-  volume    = {34},
-  number    = {2},
-  pages     = {123--131},
-  issn      = {0730-7829},
-  abstract  = {Groove is a pleasant feeling that compels people to move their bodies along with music. In the past, there was some consensus among both musicians and researchers that the main factor in inducing this feeling is onset asynchrony of sounds. However, recent studies have asserted that no-asynchrony is the condition that will obtain the highest groove. The current study examined whether no-asynchrony exclusively elicits the highest groove. In Experiment 1, we measured the groove increment of a backbeat drum pattern as a function of the asynchronies between bass guitar and hi-hat cymbal sounds. Upon evaluation, the scores of no conditions exceeded those of the synchronous condition. However, the condition with slight bass guitar precedence over the hi-hat achieved an approximately equal score to the condition with complete synchrony, and that score was higher than the ones achieved with bass delay. In Experiment 2, we measured the participants{\textquoteright} sensitivities to timing discrimination. The results confirmed that the amount of bass precedence in Experiment 1 was perceptible to the listeners. These findings suggest that complete synchronization is not always the best condition to achieve groove and that listeners prefer perceivable asynchronies in some cases.},
[…]
