### 2019-03-13 #1166 @Datseris
Perfect, I removed all unused references! (The reference you see above without DOI does not have an associated DOI to it)
