### 2019-03-13 #1166 @labarba
Do you want to add caps protection, `{MIDI}`, in the title, so it appears in upper case?
