### 2019-03-13 #1166 @Datseris
@labarba Thanks for the suggestion. I did that now in the latest commit.
