### 2019-03-14 #1166 @Datseris
@labarba Putting `{}` in the title in the markdown file didn't work so I removed it. I'd prefer to continue with the publication process instead of adding caps protection.
