### 2019-03-14 #1166 @labarba
I meant on the title of the reference we were talking about:
https://github.com/JuliaMusic/MIDI.jl/blob/610f3a26085446e74282d026e2d6359b01194225/paper/paper.bib#L235
