+++ paper/paper.bib
@@ -232,7 +232,7 @@ year = {2014}
 }
 
 @inproceedings{Raffel,
-  title={Intuitive analysis, creation and manipulation of MIDI data with pretty\_midi},
+  title={Intuitive analysis, creation and manipulation of {MIDI} data with pretty\_midi},
   author={Raffel, Colin and Ellis, Daniel PW},
   booktitle={15th International Society for Music Information Retrieval Conference Late Breaking and Demo Papers},
   pages={84--93},
