### 2019-03-14 #1166 @labarba
Congratulations, @Datseris, your JOSS paper is now published!

Sincere thanks to the editor: @Kevin-Mattheus-Moerman, and reviewers: @ssfrr, @jfsantos 🙏
