### 2019-03-14 #1166 @Datseris
I just want to be on the safe side, since doi2bib.org cannot find the DOI of this publication.

Is the following BibTeX entry the correct one?

```latex
@article{Datseris2019,
  doi = {10.21105/joss.01166},
  url = {https://doi.org/10.21105/joss.01166},
  year  = {2019},
  month = {mar},
  publisher = {The Open Journal},
  volume = {4},
  number = {35},
  pages = {1166},
  author = {George Datseris and Joel Hobson},
  title = {{MIDI}.jl: Simple and intuitive handling of MIDI data.},
  journal = {The Journal of Open Source Software}
}
```
