### 2019-03-14 #1166 @labarba
The DOI resolves:
 https://doi.org/10.21105/joss.01166
